Movement fixed and works for thomas, need to find a way in order to make it work for enemy characters Renderer now renders all game objects onto screen. Still need to work on collisions for all in game objects.

+++ core/src/za/ttd/characters/ToothDecay.java
@@ -1,15 +1,10 @@
 package za.ttd.characters;
 
-import com.badlogic.gdx.graphics.g2d.Animation;
-import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import za.ttd.characters.objects.Position;
 
-/**
- * Created by s213391244 on 7/24/2015.
- */
 public class ToothDecay extends Enemy {
 
-    public ToothDecay(Position position, int speed){
+    public ToothDecay(Position position, float speed){
         super(position, speed, "ToothDecay");
     }
 }
